+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.1.0/)  
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).  
 
+## [1.3.5] - 2025-04-23
+### Added
+- Reset and retry to switch mode failed case in criteria test. Most retry 3 times,
+  next item when all failed.
+- Delay 10ms after get raw data failed(0x08), leave time for data preparation.
+
 ## [1.3.4] - 2025-02-27
 ### Fixed
 - Add succeed message when i2c fw update success. Which will fix himax touch update